requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.24.0
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
certifi==2018.1.18
cffi==1.11.5
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.4
daphne==1.4.2
Django==2.0.3
django-bower==5.2.0
django-filter==1.1.0
djangorestframework==3.7.7
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
Markdown==2.6.11
msgpack-python==0.5.6
nexmo==2.0.0
Pillow==5.0.0
pycparser==2.18
PyJWT==1.6.0
pytz==2018.3
redis==2.10.6
requests==2.18.4
six==1.11.0
Twisted==17.9.0
txaio==2.9.0
urllib3==1.22
websocket-client==0.47.0
zope.interface==4.4.3
